feat: adds the initial structure of the base adapter
