+++ src/common/types/cdn.types.ts
@@ -1,7 +1,13 @@
 import { SuperVizSdk } from '../../services/communicator/types';
 import { FrameSize } from '../../services/video-conference-manager/types';
 
-import { DeviceEvent, MeetingEvent, RealtimeEvent, MeetingState } from './events.types';
+import {
+  DeviceEvent,
+  MeetingEvent,
+  RealtimeEvent,
+  MeetingState,
+  MeetingConnectionStatus,
+} from './events.types';
 import { SuperVizSdkOptions } from './sdk-options.types';
 
 export interface SuperVizCdn {
@@ -11,4 +17,5 @@ export interface SuperVizCdn {
   FrameSize: typeof FrameSize;
   DeviceEvent: typeof DeviceEvent;
   MeetingState: typeof MeetingState;
+  MeetingConnectionStatus: typeof MeetingConnectionStatus;
 }
