+++ src/index.ts
@@ -12,6 +12,7 @@ import ApiService from './services/api';
 import AuthService from './services/auth-service';
 import Communicator from './services/communicator';
 import { SuperVizSdk } from './services/communicator/types';
+import { AdapterMethods, DefaultAdapterOptions } from './services/integration/base-adapter/types';
 import { FrameSize } from './services/video-conference-manager/types';
 
 const validateOptions = ({ userGroup, user, roomId }: SuperVizSdkOptions) => {
@@ -59,6 +60,7 @@ if (window) {
     DeviceEvent,
     RealtimeEvent,
     MeetingState,
+    MeetingConnectionStatus,
   };
 }
 
@@ -74,4 +76,6 @@ export {
   User,
   UserGroup,
   MeetingConnectionStatus,
+  DefaultAdapterOptions,
+  AdapterMethods,
 };
