+++ src/services/integration/base-adapter/types.ts
@@ -1,3 +1,5 @@
+import { UserOn3D } from '../users/types';
+
 export interface DefaultAdapterManager {
   isAvatarsEnabled: boolean;
   isPointersEnabled: boolean;
@@ -14,4 +16,12 @@ export interface DefaultAdapterOptions {
   isFollowAvailable?: boolean;
   isGatherAvailable?: boolean;
   isGoToAvailable?: boolean;
+}
+
+export interface AdapterMethods {
+  disablePointers: () => void;
+  enablePointers: () => void;
+  enableAvatars: () => void;
+  disableAvatars: () => void;
+  getUsersOn3D: () => UserOn3D[];
 }
